pyproject.toml
[tool.poetry]
name = "portfolio-tests"
version = "0.1.0"
description = ""
authors = ["Mason Bassett <bassettmason@gmail.com>"]
readme = "README.md"
packages = [{include = "portfolio_tests"}]

[tool.poetry.dependencies]
python = "^3.11"
pyleniumio = "^1.19.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
